rutas de ususario agregadas

+++ api/routes/users.routes.js
@@ -4,7 +4,7 @@ const User = require("../models/User");
 const bcrypt = require("bcrypt");
 const { Op } = require("sequelize");
 
-const justAdmin = require("../middlewares/justAdmin");
+const { justAdmin } = require("../middlewares/justAdmin");
 const { validateToken } = require("../utils/jwt");
 const router = express.Router();
 
@@ -14,10 +14,10 @@ const router = express.Router();
 
  */
 
-router.get("/", async (req, res) => {
+router.get("/", [validateToken, justAdmin], async (req, res) => {
   try {
     const users = await User.findAll({
-      // attributes: ["username", "email", "id", "role"],
+      attributes: ["username", "email", "id", "role"],
     });
     res.json({ ok: true, users });
   } catch (error) {
@@ -26,14 +26,16 @@ router.get("/", async (req, res) => {
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", [validateToken, justAdmin], async (req, res) => {
   const { id } = req.params;
 
   try {
-    const resDB = await User.destroy({ where: { id }, limit: 1 });
+    if (id != req.payload.id) {
+      const resDB = await User.destroy({ where: { id }, limit: 1 });
+      console.log("DELETE RES", resDB);
+    }
     // const userToDelete = User.build({ id });
     // const resDB = await userToDelete.destroy();
-    console.log("DELETE RES", resDB);
     res.json({ msg: "usuario eliminado" });
   } catch (error) {
     console.log(error);
@@ -41,18 +43,18 @@ router.delete("/:id", async (req, res) => {
   }
 });
 
-router.put("/:id", async (req, res) => {
-  const id = req.params.id;
+router.put("/:id", [validateToken], async (req, res) => {
+  const {id} = req.params;
   const body = req.body;
   console.log("PAYLOAD", req.payload);
 
-  // const { id: tokenID, role } = req.payload;
+  const { id: tokenID, role } = req.payload;
 
-  // if (tokenID != id) {
-  //   return res
-  //     .status(401)
-  //     .json({ msg: "no se puede editar el perfil de otro" });
-  // }
+  if (tokenID != id) {
+    return res
+      .status(401)
+      .json({ msg: "no puedes editar el perfil de otro" });
+  }
 
   try {
     const user = await User.findByPk(id);
@@ -71,9 +73,9 @@ router.put("/:id", async (req, res) => {
         const userDB = await User.findOne({
           where: {
             email: body.email,
-            id: {
-              [Op.not]: [id],
-            },
+            // id: {
+            //   [Op.not]: [id],
+            // },
           },
         });
 
@@ -88,8 +90,7 @@ router.put("/:id", async (req, res) => {
       const [, userUpdated] = await User.update(body, {
         where: { id },
         fields: ["username", "password", "email"],
-        // returning: ["id", "role", "email", "username"],
-        returning: true,
+        returning: ["id", "role", "email", "username"],
       });
 
       res.json({ ok: true, msg: "el usuario fue actualizado", userUpdated });
